Dev final release(mk)

+++ src/components/projects/ProjectDetailContainer.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import LoaderCircular from './../common/LoaderCircular';
 import Icon from './../common/Icon';
-import DCMTWebApi from '../../api';
+import WebApi from '../../api';
 import { withRouter } from 'react-router';
 import SkillRow from './../skills/SkillRow';
 import Modal from 'react-responsive-modal';
 import SkillsSelect from './../skills/SkillsSelect';
 import DetailCascade from './../employees/DetailCascade';
 import ProjectOwner from './ProjectOwner';
-import { setActionConfirmation } from './../../actions/asyncActions';
+import { setActionConfirmation, setActionConfirmationProgress, setActionConfirmationResult } from './../../actions/asyncActions';
 import { connect } from 'react-redux';
 import { translate } from 'react-translate';
 import EditProjectDetails from './modals/EditProjectDetails';
@@ -16,7 +16,12 @@ import { bindActionCreators } from 'redux';
 import * as projectsActions from "../../actions/projectsActions";
 import TeamMember from './TeamMember';
 import AddProjectOwner from './modals/AddProjectOwner';
-import { SET_ACTION_CONFIRMATION_RESULT } from '../../constants';
+import { SET_ACTION_CONFIRMATION_RESULT, ACTION_CONFIRMED } from '../../constants';
+import AddEmployeeToProject from '../employees/modals/AddEmployeeToProject';
+import AssignmentModal from './../assign/AssignmentModal';
+import binaryPermissioner from './../../api/binaryPermissioner';
+import specialPermissioner from './../../api/specialPermissioner';
+import IntermediateBlock from './../common/IntermediateBlock';
 
 class ProjectDetailContainer extends Component {
   constructor(props) {
@@ -34,14 +39,18 @@ class ProjectDetailContainer extends Component {
       seniorityLevel: 1,
       project: {},
       showEditProjectModal: false,
+      showAddEmployee: false,
+      addEmployeeStageTwo: false,
+      addEmployeeSelectionId: {},
       pps_rb: {},
       showAddOwner: false,
-      projectActions: bindActionCreators(projectsActions, this.props.dispatch)
+      projectActions: bindActionCreators(projectsActions, this.props.dispatch),
+      rowUnfurls: {}
     };
   }
 
   componentDidMount() {
-    this.getProject(this.props.match.params.id);
+    this.refresh();
   }
 
   componentWillReceiveProps(nextProps) {
@@ -51,31 +60,59 @@ class ProjectDetailContainer extends Component {
       'reactivateProject',
       'deleteProject'
     ];
+    if (this.validatePropsForAction(nextProps, "deleteProjectMember")) {
+      this.props.dispatch(setActionConfirmationProgress(true));
+      const { assignmentId } = this.props.toConfirm;
+      WebApi.assignments.delete(assignmentId)
+        .then(response => {
+          this.props.dispatch(setActionConfirmationResult(response));
+          this.refresh();
+        })
+        .catch(error => {
+          this.props.dispatch(setActionConfirmationResult(error));
+        });
+    }
     if(nextProps.type === SET_ACTION_CONFIRMATION_RESULT){
       if(acceptableKeys.indexOf(this.props.toConfirm.key) >= 0){
         setTimeout(() => {
-          window.location.reload();
-        }, 500);
+          this.refresh();
+        }, 400);
       }
     }
   }
 
+  refresh = () => {
+    this.getProject(this.props.match.params.id);
+  }
+
+  validatePropsForAction(nextProps, action) {
+    return (
+      nextProps.confirmed &&
+      !nextProps.isWorking &&
+      nextProps.type === ACTION_CONFIRMED &&
+      nextProps.toConfirm.key === action
+    );
+  }
+
   getProject = (id) => {
-    DCMTWebApi.getProject(id)
+    WebApi.projects.get(id)
       .then((response) => {
-        console.log('PROJECT', response.data.dtoObject);
+        let extract = response.extractData();
         this.setState({
-          project: response.data.dtoObject,
+          project: extract,
           loading: false,
           projectLoadedSuccessfully: true,
-          projectActive: response.data.dtoObject.isActive,
-          skills: response.data.dtoObject.skills,
-          team: (response.data.dtoObject.team !== undefined ?
-            response.data.dtoObject.team : [])
+          projectActive: extract.isActive,
+          skills: extract.skills,
+          team: (extract.team !== undefined ?
+            extract.team : []),
+          replyBlock: response
         });
       })
       .catch((error) => {
-
+        this.setState({
+          replyBlock: error
+        });
       });
   }
 
@@ -98,13 +135,11 @@ class ProjectDetailContainer extends Component {
     this.setState({
       loading: true
     });
-    DCMTWebApi.putProjectSkills(this.props.match.params.id, this.state.skills)
+    WebApi.projects.put.skills(this.props.match.params.id, this.state.skills)
       .then((response) => {
         this.setState({
           loading: false,
-          pps_rb: {
-            response
-          },
+          pps_rb: response,
           edit: false
         });
       })
@@ -199,7 +234,7 @@ class ProjectDetailContainer extends Component {
 
   handleEditProjectCloseModal = (success = false) => {
     this.setState({ showEditProjectModal: false }, () => {
-      if(success) window.location.reload();
+      if(success) this.refresh();
     });
   }
 
@@ -240,10 +275,16 @@ class ProjectDetailContainer extends Component {
 
   handleCloseAddOwner = (success = false) => {
     this.setState({ showAddOwner: false }, () => {
-      if(success) setTimeout(() => {
-        window.location.reload();
-      });
+      this.refresh();
     });
+  }
+
+  handleOpenAddEmployee = () => {
+    this.setState({ showAddEmployee: true, addEmployeeStageTwo: false, addEmployeeSelectionId: "" });
+  }
+
+  handleCloseAddEmployee = (success = false) => {
+    this.setState({ showAddEmployee: false, addEmployeeStageTwo: false });
   }
 
   mapSkills = (skills, editable = false) => {
@@ -260,24 +301,73 @@ class ProjectDetailContainer extends Component {
     });
   }
 
+  handleRowClick = (object, index) => {
+    return (e) => {
+      const { rowUnfurls } = this.state;
+
+      if(rowUnfurls[index] === undefined){
+        rowUnfurls[index] = true;
+      } else {
+        rowUnfurls[index] = !rowUnfurls[index];
+      }
+
+      this.setState({
+        rowUnfurls
+      });
+    };
+  }
+
+  deleteMember = (assignment) => (e) => {
+    e.stopPropagation();
+    e.nativeEvent.stopImmediatePropagation();
+
+    this.props.dispatch(
+      setActionConfirmation(true, {
+        key: "deleteProjectMember",
+        string: `Chcesz wypisać ${assignment.firstName} ${assignment.lastName} z projektu ${this.state.project.name}`,
+        assignmentId: assignment.assignmentId,
+        successMessage: "Wypisano pracownika"
+      })
+    );
+  }
+
   mapTeam = (team) => {
     return team.map((teamAssignment, index) => {
+      let unfurled = this.state.rowUnfurls[index];
       return (
-        <TeamMember key={index} assignment={teamAssignment}/>
+        [
+          <TeamMember
+            onClick={this.handleRowClick(teamAssignment, index)}
+            compact
+            key={index}
+            assignment={teamAssignment}
+            delete={this.deleteMember}
+          />,
+          unfurled ? <TeamMember
+            onClick={this.handleRowClick(teamAssignment, index)}
+            key={50000 - index}
+            assignment={teamAssignment}
+          /> : null
+        ]
       );
     });
   }
 
   mapOwners = (owners, project) => {
+    let allowed = specialPermissioner().projects.isOwner(project, this.props.login)
+    || binaryPermissioner(false)(0)(0)(0)(0)(1)(1)(this.props.binPem);
     return [
       ... owners.map((owner, index) => {
-        return <ProjectOwner clickAction={this.deleteProjectOwner(owner, project)} key={index} owner={owner}/>;
+        if(allowed){
+          return <ProjectOwner clickAction={this.deleteProjectOwner(owner, project)} key={index} owner={owner}/>;
+        }
+        else return <ProjectOwner key={index} owner={owner}/>;
       }),
-      <div key={-1} className="project-owner">
+      allowed ? <div key={-1} className="project-owner">
         <span onClick={this.handleOpenAddOwner} className="project-owner-add">
           <span/>
         </span>
-      </div>
+      </div> : null
     ];
   }
 
@@ -321,6 +411,14 @@ class ProjectDetailContainer extends Component {
     );
   }
 
+  addEmployeeStageTwo = (employee) => {
+    console.log(employee);
+    this.setState({
+      addEmployeeStageTwo: true,
+      addEmployeeSelection: employee
+    });
+  }
+
   pullProjectEditModalDOM = () => {
     return <Modal
       open={this.state.showEditProjectModal}
@@ -352,6 +450,31 @@ class ProjectDetailContainer extends Component {
     </Modal>;
   }
 
+  pullAddEmployeeModalDOM = () => {
+    return <Modal
+      open={this.state.showAddEmployee}
+      classNames={{ modal: "Modal Modal-add-employee Modal-assignment" }}
+      contentLabel="Add employee modal"
+      onClose={this.handleCloseAddEmployee}
+    >
+      {
+        this.state.addEmployeeStageTwo ?
+        <AssignmentModal
+          employee={this.state.addEmployeeSelection}
+          project={this.state.project}
+          refresh={this.refresh}
+        />
+        :
+        <AddEmployeeToProject
+          aetpStageTwo={this.addEmployeeStageTwo}
+          project={this.state.project}
+          completed={this.handleOwnerSelectionFinale}
+        />
+      }
+
+    </Modal>;
+  }
+
   pullModalDOM = () => {
     return <Modal
         open={this.state.showModal}
@@ -366,7 +489,7 @@ class ProjectDetailContainer extends Component {
   pullEmployeeIdBlockDOM = () => {
     const { t } = this.props;
     const { project } = this.state;
-    return <div className="col-sm-4 project-id-block button-group">
+    return <div className="col-xl-4 col-sm-12 project-id-block button-group">
       <Icon icon="briefcase" iconSize="2x"/>
       <h1>{project.name}</h1>
       {
@@ -375,16 +498,24 @@ class ProjectDetailContainer extends Component {
         : <h3 className="project-inactive">{t("Inactive")}</h3>
       }
       <hr className="sharp"/>
-      <button onClick={this.changeSettings} className="project-headway dcmt-button button-success">{t("EditProject")}</button>
       {
-        this.state.project.isActive ?
-        <button onClick={this.closeProject} title={t("Close")} className="project-headway dcmt-button button-lowkey">{t("Close")}</button>
-        : <button onClick={this.reactivateProject} title={t("Reactivate")} className="project-headway dcmt-button button-success">{t("Reactivate")}</button>
+        specialPermissioner().projects.isOwner(project, this.props.login)
+        || binaryPermissioner(false)(0)(0)(0)(0)(1)(1)(this.props.binPem) ?
+        <button onClick={this.changeSettings} className="project-headway dcmt-button button-success">{t("EditProject")}</button>
+        : null
       }
       {
-        this.state.project.isDeleted ?
-        null
+        (specialPermissioner().projects.isOwner(project, this.props.login)
+        || binaryPermissioner(false)(0)(0)(0)(0)(1)(1)(this.props.binPem)) ? this.state.project.isActive ?
+        <button onClick={this.closeProject} title={t("Close")} className="project-headway dcmt-button button-lowkey">{t("Close")}</button>
+        : <button onClick={this.reactivateProject} title={t("Reactivate")} className="project-headway dcmt-button button-success">{t("Reactivate")}</button>
+        : null
+      }
+      {
+        (specialPermissioner().projects.isOwner(project, this.props.login)
+        || binaryPermissioner(false)(0)(0)(0)(0)(1)(1)(this.props.binPem)) ? this.state.project.isDeleted ? null
         : <button onClick={this.deleteProject} title={t("Delete")} className="project-headway dcmt-button button-lowkey">{t("Delete")}</button>
+        : null
       }
       <hr/>
       <div className="project-headway">
@@ -392,8 +523,8 @@ class ProjectDetailContainer extends Component {
         <DetailCascade lKey={t("Client")} rVal={project.client} lColSize={4} rColSize={8} />
         <DetailCascade lKey={t("Deleted")} rVal={project.isDeleted ? t("Yes") : t("No")} lColSize={4} rColSize={8} />
         <DetailCascade lKey={t("Active")} rVal={project.isActive ? t("Yes") : t("No")} lColSize={4} rColSize={8} />
-        <DetailCascade lKey={t("StartDate")} rVal={project.startDate} lColSize={4} rColSize={8} />
-        <DetailCascade lKey={t("EstimatedEndDate")} rVal={project.estimatedEndDate} lColSize={4} rColSize={8} />
+        <DetailCascade lKey={t("StartDate")} rVal={(new Date(project.startDate).toLocaleDateString())} lColSize={4} rColSize={8} />
+        <DetailCascade lKey={t("EstimatedEndDate")} rVal={(new Date(project.estimatedEndDate).toLocaleDateString())} lColSize={4} rColSize={8} />
       </div>
       <hr/>
       <div className="project-headway project-text-justified">
@@ -442,20 +573,60 @@ class ProjectDetailContainer extends Component {
     const { t } = this.props;
     return <div className="row">
       { this.state.projectLoadedSuccessfully ? this.pullEmployeeIdBlockDOM() : null }
-      <div className="col-sm-7 project-headway">
-        {this.mapTeam(this.state.team)}
+      <div className="col-xl-8 col-sm-12 project-headway">
+        <div className="row">
+          <div className="col-sm-12">
+            <div className="project-headway project-bold">{'Zespół'}</div>
+            <table className="team-member-compact-table">
+              <thead>
+                <tr>
+                  <th></th>
+                  <th>Name</th>
+                  <th>Capacity</th>
+                  <th>Role</th>
+                  <th>Seniority</th>
+                  <th>Position</th>
+                  <th>Ends</th>
+                  <th>!</th>
+                </tr>
+              </thead>
+              <tbody>
+                {this.mapTeam(this.state.team)}
+              </tbody>
+            </table>
+          </div>
+        </div>
+        <div className="row">
+          <div className="col-lg-10"/>
+          <div className="col-lg-2 full-width-button">
+            {
+            binaryPermissioner(false)(0)(0)(1)(1)(1)(1)(this.props.binPem) ?
+            <button onClick={this.handleOpenAddEmployee} className="dcmt-button button-success project-headway">
+              {'Dodaj'}
+            </button> : null
+            }
+          </div>
+        </div>
         <hr/>
-        {this.mapSkills(this.state.skills)}
-      </div>
-      <div className="col-sm-1 project-headway full-width-button">
-        {
-          this.state.edit === false ?
-          <button onClick={this.edit} className="dcmt-button">
-            {t("Edit")}
-          </button>
-          :
-          this.pullEditToolbarDOM()
-        }
+        <div className="project-headway project-bold">{'Umiejętności powiązane'}</div>
+        <div className="row">
+          <div className="col-xl-10 col-sm-12">
+            {this.mapSkills(this.state.skills)}
+          </div>
+          <div className="col-xl-2 col-sm-12 full-width-button">
+          {
+            (specialPermissioner().projects.isOwner(project, this.props.login)
+            || binaryPermissioner(false)(0)(0)(0)(0)(1)(1)(this.props.binPem)) ? this.state.edit === false ?
+            <button onClick={this.edit} className="dcmt-button">
+              {t("Edit")}
+            </button>
+            :
+            this.pullEditToolbarDOM()
+            :
+            null
+          }
+          </div>
+        </div>
       </div>
     </div>;
   };
@@ -466,7 +637,14 @@ class ProjectDetailContainer extends Component {
         { this.pullModalDOM() }
         { this.pullProjectEditModalDOM() }
         { this.pullAddOwnerModalDOM() }
-        { this.state.loading ? <LoaderCircular/> : this.pullDOM() }
+        { this.pullAddEmployeeModalDOM() }
+        {
+          <IntermediateBlock
+            loaded={!this.state.loading}
+            render={this.pullDOM}
+            resultBlock={this.props.replyBlock}
+          />
+        }
       </div>
     );
   }
@@ -479,7 +657,9 @@ function mapStateToProps(state) {
     toConfirm: state.asyncReducer.toConfirm,
     isWorking: state.asyncReducer.isWorking,
     resultBlock: state.asyncReducer.resultBlock,
-    type: state.asyncReducer.type
+    type: state.asyncReducer.type,
+    binPem: state.authReducer.binPem,
+    login: state.authReducer.login
   };
 }
